requirements.txt
Django>=3.0,<=3.1
whitenoise
psycopg2-binary
djangorestframework
django-cors-headers
django-storages[azure]
django-rest-auth
django-allauth
numpy
pillow